requirements.txt
streamlit
openai
# playwright
asyncio
streamlit
seleniumbase
webdriver-manager